Add GitHub search option/feature
This is somewhat of a rough PoC for enabling the use of GitHub repo search, rather than relying on a list of all repos accessible to the user. In the future, there might be a more user-friendly way to do this.

Anyways, this adds two new GitHub-related env vars:
- `GITHUB_USE_SEARCH`: if `true`, uses octonode's `search.reposAsync` instead of `reposAsync` passing in a search query
- `GITHUB_SEARCH_QUERY`: specifies the search query to use when searching for repos for Threat Dragon to use

A conditional check in `threatmodelcontroller` determines whether to use `reposAsync` _or_ `search.reposAsync` based on the above.

Finally, this adds a wrapper function in `threatmodelrepository` to use octonode's `search.reposAsync`, passing in the aforementioned search query.